couchdb: Explicitly fail when no database name
Nano throws a rather cryptic message when attempting to create database
which does not exist related to invalid parameters.  Adding more
confusion nano uses a single exception which is created on module load,
thus making the stack trace useless when trying to diagnose the problem.

If a falsy name is passed to the `couchdb.Service#ensure_db_exists
method it will now throw an assertion error.